requirements.txt
absl-py==2.1.0
addict==2.4.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
basicsr==1.4.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==4.2.3
executing==2.0.1
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
gfpgan==1.3.8
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
jax==0.4.25
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.12.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml-dtypes==0.3.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==1.5.3
parso==0.8.3
Pillow==9.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
RealESRGAN @ git+https://github.com/sberbank-ai/Real-ESRGAN.git@362a0316878f41dbdfbb23657b450c3353de5acf
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tb-nightly==2.17.0a20240324
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.2.12
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
yapf==0.40.2
zipp==3.18.1
